pyproject.toml
[project]
name = "brasa-backend"
requires-python = ">=3.10"
authors = [{ name="Sergio Pires", email="sergiodanpires@gmail.com" }]
description = "A simple backend for brasa app"
readme = "README.md"
license = { file = "LICENSE" }
classifiers = [
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12"
]
dynamic = ["version"]
dependencies = [
  "flask>=3.0.3",
  "flask>=1.3.0",
  "pymongo>=4.7.2",
  "dataclasses_json>=0.6.6",
  "requests==2.31.0",
  "redis==5.0.4",
  "requests==2.31.0",
  "pytz==2021.3",
  "python-dotenv==1.0.0",
  "prometheus_client==0.20.0"
]

[project.urls]
repository = "https://github.com/Sergio-Daniel-Pires/brasa-backend"

[project.optional-dependencies]
test = [
  "pytest==7.4.2"
]
